Fixed printf type warning on macOS.

+++ wpad_dhcp_posix.c
@@ -179,7 +179,7 @@ static bool dhcp_read_reply(SOCKET sfd, uint32_t request_xid, dhcp_msg *reply) {
     const ssize_t response_len = recvfrom(sfd, (char *)reply, sizeof(dhcp_msg), 0, NULL, NULL);
 
     if (response_len <= (ssize_t)(sizeof(dhcp_msg) - DHCP_OPT_MIN_LENGTH)) {
-        LOG_DEBUG("Unable to read DHCP reply (%d:%d)\n", response_len, socketerr);
+        LOG_DEBUG("Unable to read DHCP reply (%d:%d)\n", (int32_t)response_len, socketerr);
         return false;
     }
